moved hegde graphs to their own crate
