+++ Cargo.toml
@@ -44,6 +44,8 @@ criterion = { version = "0.5", features = ["html_reports"] }
 # These will be patched
 symbolica = { version = "0.14.0" }
 spenso = { version = "0.4.0", features = ["shadowing"] }
+linnet = {git =  "https://github.com/alphal00p/linnet",features = ["symbolica","drawing"]}
+
 
 momtrop = { git = "https://github.com/alphal00p/momtrop", features = ["log"] }
 serde = { version = "1.0", features = ["derive"] }
